requirements.txt
Appium-Python-Client==2.11.1
selenium==4.11.2
opencv-python==4.8.0.76
numpy==1.24.3
pillow==10.0.0
pytest==7.4.0
webdriver-manager==4.0.0
requests==2.31.0
python-dotenv==1.0.0
loguru==0.7.0
pandas==2.0.3
matplotlib==3.7.2 